+++ libraries/classes/Controllers/Table/Structure/AbstractIndexController.php
@@ -1,0 +1,74 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpMyAdmin\Controllers\Table\Structure;
+
+use PhpMyAdmin\Controllers\AbstractController;
+use PhpMyAdmin\Controllers\Table\StructureController;
+use PhpMyAdmin\DatabaseInterface;
+use PhpMyAdmin\Http\ServerRequest;
+use PhpMyAdmin\Message;
+use PhpMyAdmin\ResponseRenderer;
+use PhpMyAdmin\Template;
+use PhpMyAdmin\Util;
+
+use function __;
+use function count;
+use function is_array;
+
+abstract class AbstractIndexController extends AbstractController
+{
+    /** @var DatabaseInterface */
+    private $dbi;
+
+    /** @var StructureController */
+    private $structureController;
+
+    public function __construct(
+        ResponseRenderer $response,
+        Template $template,
+        DatabaseInterface $dbi,
+        StructureController $structureController
+    ) {
+        parent::__construct($response, $template);
+        $this->dbi = $dbi;
+        $this->structureController = $structureController;
+    }
+
+    public function handleIndexCreation(ServerRequest $request, string $indexType): void
+    {
+        $GLOBALS['message'] = $GLOBALS['message'] ?? null;
+
+        $selected = $request->getParsedBodyParam('selected_fld', []);
+
+        if (! is_array($selected) || $selected === []) {
+            $this->response->setRequestStatus(false);
+            $this->response->addJSON('message', __('No column selected.'));
+
+            return;
+        }
+
+        $i = 1;
+        $selectedCount = count($selected);
+        $GLOBALS['sql_query'] = 'ALTER TABLE ' . Util::backquote($GLOBALS['table']) . ' ADD ' . $indexType . '(';
+
+        foreach ($selected as $field) {
+            $GLOBALS['sql_query'] .= Util::backquote($field);
+            $GLOBALS['sql_query'] .= $i++ === $selectedCount ? ');' : ', ';
+        }
+
+        $this->dbi->selectDb($GLOBALS['db']);
+        $result = $this->dbi->tryQuery($GLOBALS['sql_query']);
+
+        if (! $result) {
+            $GLOBALS['message'] = Message::error($this->dbi->getError());
+        }
+
+        if (empty($GLOBALS['message'])) {
+            $GLOBALS['message'] = Message::success();
+        }
+
+        ($this->structureController)($request);
+    }
+}
